Change Lore-Master
Identifying all items with the mpa changes to Lore-Keeper and the id
game was pretty lackluster. Now there's no way out of the id game,
Lore-Master does not identify all items.

Lore-Master now allows the player to perceive a level feeling like in
Angband, though there are only three levels of feeling:
 - Mundane
 - Forge or lesser vault
 - Enchanted forge or greater vault
The feeling is not revealed until a perception vs. staircasiness/100
check is won, as a brake on scumming.

The feeling is revealed through a message, can be reviewed through ^F,
and the depth is coloured yellow or red for lesser or greater feeling
(parallel with enchanted and unique forge colours).

To not be obnoxious, feeling information continues to be available
through the end of the level even if the player loses Lore-Master, since
the cat is out of the bag.