Remove the no_light function.
This checked if the player's grid is dark, but seemed to be used as a
proxy for giving penalties when a player couldn't see something they
were interacting with. These penalties didn't stack with blindness.

The following actions interact with no_light
  - searching
  - opening chests
  - disarming chests
  - unlocking doors
  - disarming traps

These all now give a penalty if the grid is not seen for any reason.

I'm considering giving an extra penalty to a player in the dark
attempting these actions (you can't see your own hands), considering the
balance changes of the previous commit. This would only have an effect
in the presence of darkness generating monsters or even stranger edge
cases, though.